rename utils to test_utils
py.test only does assertion introspection in files named "test_*.py",
so asserts that take place in other files such as utils.py do not
benefit from pytest's nice asserts where it prints out values passed in
and such, so all files containing `assert x` lines should be prefixed
with `test_`. utils.py doesn't have any functions that start with
test_, so won't affect how py.test runs
