+++ tests/notify_api/test_notify_api_email.py
@@ -6,7 +6,7 @@ from tests.postman import (
     get_notification_by_id_via_api
 )
 
-from tests.utils import assert_notification_body, recordtime
+from tests.test_utils import assert_notification_body, recordtime
 
 
 @recordtime
